requirements.txt
pandas
selenium
requests2
lxml